surrogate-lab/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "surrogatelab"
version = "0.1.0"
description = "Modular ML pipeline for FEBio simulation surrogate modeling"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "torch>=2.0.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "mlflow>=2.10.0",
    "pyyaml>=6.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
sl-train = "scripts.train:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "scripts/*", "notebooks/*"]
